ulux-binding: Move command / update handling to their own class
